Refactor Redis configuration to use structured environment variables
- Replace REDIS_URL with REDIS_HOST, REDIS_PORT, REDIS_PASSWORD
- Update redis_cache.py to use individual connection parameters
- Update celery_app.py to build Redis URL from components
- Move _execute_task_with_celery from examples.py to executor.py
- Update all .env files and Docker Compose configurations
- Add Redis configuration section to documentation

Benefits:
- Clearer configuration with separate host/port parameters
- Better Docker compatibility with service names
- Consistent URL building across all modules
- Proper placement of core utility functions
