+++ .env.dev
@@ -22,7 +22,8 @@ POSTGRES_DB=codeability
 POSTGRES_PORT=5432
 POSTGRES_URL=${POSTGRES_HOST}:${POSTGRES_PORT}
 
-REDIS_URL=localhost
+REDIS_HOST=localhost
+REDIS_PORT=6379
 REDIS_PASSWORD=redis_password
 
 DEBUG_MODE=development
