+++ .env.prod
@@ -23,7 +23,8 @@ POSTGRES_DB=codeability
 POSTGRES_PORT=5437
 POSTGRES_URL=${POSTGRES_HOST}:${POSTGRES_PORT}
 
-REDIS_URL=172.17.0.1
+REDIS_HOST=172.17.0.1
+REDIS_PORT=6379
 REDIS_PASSWORD=redis_password
 
 DEBUG_MODE=production
