+++ src/ctutor_backend/redis_cache.py
@@ -1,11 +1,16 @@
 import os
 from aiocache import Cache
 
+# Get Redis configuration from environment
+REDIS_HOST = os.environ.get('REDIS_HOST', 'localhost')
+REDIS_PORT = int(os.environ.get('REDIS_PORT', '6379'))
+REDIS_PASSWORD = os.environ.get('REDIS_PASSWORD', '')
+
 _redis_cache = Cache(
     Cache.REDIS,
-    endpoint=os.environ.get("REDIS_URL"),
-    port=6379,
-    password=os.environ.get("REDIS_PASSWORD"),
+    endpoint=REDIS_HOST,
+    port=REDIS_PORT,
+    password=REDIS_PASSWORD if REDIS_PASSWORD else None,
     pool_max_size=10,
     db=0
 )
