+++ src/ctutor_backend/tasks/examples.py
@@ -5,11 +5,11 @@ Example task implementations demonstrating the task execution framework.
 import asyncio
 from typing import Any, Dict
 from datetime import datetime
-from celery import states
 
 from .base import BaseTask
 from .registry import register_task
 from .celery_app import app
+from .executor import _execute_task_with_celery
 
 
 @register_task
@@ -171,77 +171,3 @@ def example_failing_celery_task(self, **kwargs):
     return _execute_task_with_celery(self, ExampleFailingTask, **kwargs)
 
 
-def _execute_task_with_celery(celery_task, task_class, **kwargs):
-    """
-    Execute a BaseTask implementation within Celery.
-    
-    Args:
-        celery_task: The Celery task instance (self)
-        task_class: The BaseTask class to execute
-        **kwargs: Task parameters
-        
-    Returns:
-        Task execution result
-    """
-    # Create task instance
-    task_instance = task_class()
-    
-    # Update task state to STARTED
-    celery_task.update_state(
-        state=states.STARTED,
-        meta={
-            'started_at': datetime.utcnow().isoformat(),
-            'progress': {'status': 'started'}
-        }
-    )
-    
-    try:
-        # Execute task in asyncio event loop
-        import asyncio
-        try:
-            loop = asyncio.get_event_loop()
-        except RuntimeError:
-            loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(loop)
-        
-        try:
-            # Execute the task
-            result = loop.run_until_complete(task_instance.execute(**kwargs))
-            
-            # Call success hook
-            loop.run_until_complete(task_instance.on_success(result, **kwargs))
-            
-            # Update final state
-            celery_task.update_state(
-                state=states.SUCCESS,
-                meta={
-                    'started_at': datetime.utcnow().isoformat(),
-                    'finished_at': datetime.utcnow().isoformat(),
-                    'progress': {'status': 'completed'}
-                }
-            )
-            
-            return result
-            
-        except Exception as e:
-            # Call failure hook
-            try:
-                loop.run_until_complete(task_instance.on_failure(e, **kwargs))
-            except Exception:
-                pass  # Don't let failure hook errors mask the original error
-            
-            # Re-raise the original exception
-            raise e
-            
-    except Exception as e:
-        # Update error state
-        celery_task.update_state(
-            state=states.FAILURE,
-            meta={
-                'started_at': datetime.utcnow().isoformat(),
-                'finished_at': datetime.utcnow().isoformat(),
-                'error': str(e),
-                'progress': {'status': 'failed'}
-            }
-        )
-        raise e
